requirements.txt
python 3.8
pytorch 1.12.1  (cudatoolkit 11.6)
transformers 4.27.4
datasets 2.11.0
sentence-transformers 2.2.2
sentencepiece 0.1.97
language-tool-python 2.7.1
protobuf 3.19.0
sklearn-pandas 2.2.0
python-Levenshtein 0.20.9


